requirements.txt
boto3
qdrant_client
python-dotenv
pymupdf
openai
pdfplumber
openpyxl
pandas
numpy
tabulate
langchain
langchain-text-splitters